requirements.txt
python-telegram-bot
redis
groq
python-dotenv
gspread
google-auth-oauthlib
